Explicitly convert to float in JoltCustomChunkShape::GetInnerRadius

+++ src/shapes/jolt_chunk_shape_impl_3d.hpp
@@ -18,7 +18,7 @@ public:
 
 	virtual JPH::uint GetSubShapeIDBitsRecursive() const override { return 0; }
 
-	virtual float GetInnerRadius() const override { return size * 0.5f; }
+	virtual float GetInnerRadius() const override { return (float) size * 0.5f; }
 
 	virtual JPH::MassProperties GetMassProperties() const override {
 		JPH::MassProperties mass_properties;
